# [Performance] Speed up Azure A10 instance creation