--- a/sky/clouds/service_catalog/images/skypilot-aws-cpu-ubuntu.pkr.hcl
+++ b/sky/clouds/service_catalog/images/skypilot-aws-cpu-ubuntu.pkr.hcl
@@ -4,11 +4,11 @@ variable "region" {
 }
 
 locals {
-  timestamp = regex_replace(timestamp(), "[- TZ:]", "")
+  date = formatdate("YYMMDD", timestamp())
 }
 
 source "amazon-ebs" "cpu-ubuntu" {
-  ami_name      = "skypilot-aws-cpu-ubuntu-${local.timestamp}"
+  ami_name      = "skypilot-aws-cpu-ubuntu-${local.date}"
   instance_type = "t2.micro"
   region        = var.region
   ssh_username  = "ubuntu"