--- a/sky/clouds/service_catalog/images/skypilot-azure-cpu-ubuntu.pkr.hcl
+++ b/sky/clouds/service_catalog/images/skypilot-azure-cpu-ubuntu.pkr.hcl
@@ -15,6 +15,7 @@ locals {
 
 source "azure-arm" "cpu-ubuntu" {
   managed_image_resource_group_name = "skypilot-images"
+  // TODO(yika): these fields may not be required as we use community images below instead. We need to double-check if these can be removed.
   managed_image_name                = "skypilot-azure-cpu-ubuntu-${local.date}"
 
   subscription_id = "59d8c23c-7ef5-42c7-b2f3-a919ad8026a7"