--- a/sky/clouds/service_catalog/images/skypilot-azure-gpu-ubuntu.pkr.hcl
+++ b/sky/clouds/service_catalog/images/skypilot-azure-gpu-ubuntu.pkr.hcl
@@ -8,6 +8,12 @@ variable "vm_generation" {
   description = "Azure's VM generation, currently support 1 or 2"
 }
 
+variable "use_grid_driver" {
+  type        = bool
+  default     = false
+  description = "Whether to use the Azure GRID driver. Currently only A10 GPU VMs need this."
+}
+
 locals {
   date = formatdate("YYMMDD", timestamp())
   version   = formatdate("YY.MM.DD", timestamp())
@@ -26,8 +32,8 @@ source "azure-arm" "gpu-ubuntu" {
   image_publisher = "Canonical"
   image_offer     = "0001-com-ubuntu-server-jammy"
   image_sku       = var.vm_generation == 1 ? "22_04-lts" : "22_04-lts-gen2"
-  location        = var.vm_generation == 1 ? "eastus" : "centralus"
-  vm_size         = var.vm_generation == 1 ? "Standard_NC4as_T4_v3" : "Standard_NC24ads_A100_v4"
+  location        = var.use_grid_driver || var.vm_generation == 1 ? "eastus" : "centralus"
+  vm_size         = var.use_grid_driver ? "Standard_NV36ads_A10_v5" : (var.vm_generation == 1 ? "Standard_NC4as_T4_v3" : "Standard_NC24ads_A100_v4")
   ssh_username    = "azureuser"
   azure_tags = {
     Created_by = "packer"
@@ -37,8 +43,8 @@ source "azure-arm" "gpu-ubuntu" {
   shared_image_gallery_destination {
     subscription   = "59d8c23c-7ef5-42c7-b2f3-a919ad8026a7"
     resource_group = "skypilot-images"
-    gallery_name   = var.vm_generation == 1 ? "skypilot_images": "skypilot_image_gallery"
-    image_name     = "skypilot-gpu-gen${var.vm_generation}"
+    gallery_name   = var.use_grid_driver || var.vm_generation == 1 ? "skypilot_images" : "skypilot_image_gallery"
+    image_name     = var.use_grid_driver ? "skypilot-gpu-gen2-grid" : "skypilot-gpu-gen${var.vm_generation}"
     image_version  = "${local.version}"
     replication_regions = [
       "centralus",
@@ -61,7 +67,7 @@ build {
     script = "./provisioners/docker.sh"
   }
   provisioner "shell" {
-    script = "./provisioners/cuda.sh"
+    script = var.use_grid_driver ? "./provisioners/cuda-azure-grid.sh" : "./provisioners/cuda.sh"
   }
   provisioner "shell" {
     script = "./provisioners/nvidia-container-toolkit.sh"