--- a/sky/clouds/service_catalog/images/skypilot-gcp-gpu-ubuntu.pkr.hcl
+++ b/sky/clouds/service_catalog/images/skypilot-gcp-gpu-ubuntu.pkr.hcl
@@ -4,11 +4,11 @@ variable "zone" {
 }
 
 locals {
-  timestamp = regex_replace(timestamp(), "[- TZ:]", "")
+  date = formatdate("YYMMDD", timestamp())
 }
 
 source "googlecompute" "gpu-ubuntu" {
-  image_name          = "skypilot-gcp-gpu-ubuntu-${local.timestamp}"
+  image_name          = "skypilot-gcp-gpu-ubuntu-${local.date}"
   project_id          = "sky-dev-465"
   source_image_family = "ubuntu-2204-lts"
   zone                = var.zone