<!-- Describe the changes in this PR -->
Use the new custom image for Azure's A10 VM creation. 

Example performance change (with fixed region):
|  VM Type 💻   | Old Provision 🕐                 | New Provision 🕐     | % Speedup ✅ |
| --------------- | ---------------------------- | ----------- | ----------- |
| Azure GPU      |   21min    | 2min 50s   | 85% (7x) |

<!-- Describe the tests ran -->
<!-- Unit tests (tests/test_*.py) are part of GitHub CI; below are tests that launch on the cloud. -->

Tested (run the relevant ones):

- [x] Code formatting: `bash format.sh`
- [x] Manual test: `time sky launch --cloud=azure -y --gpus=A10:1` and ssh to machine to confirm NVIDIA and nvidia-docker runtime installation