@Michaelvll 2024-10-30
A future TODO: we need to add a smoke test for A10 on Azure